+++ packages/composites/google-weather/src/endpoint/geojson-validation.d.ts
@@ -1,3 +1,0 @@
-declare module 'geojson-validation'
-
-declare function valid(geoJSONObject: GeoJSON, trace = false): boolean