+++ packages/composites/implied-price/src/endpoint/computedPrice.ts
@@ -1,14 +1,12 @@
-import type {
+import {
+  AdapterError,
+  AdapterInputError,
   AdapterRequest,
+  AdapterResponseInvalidError,
   AxiosRequestConfig,
   Config,
   ExecuteWithConfig,
   InputParameters,
-} from '@chainlink/ea-bootstrap'
-import {
-  AdapterError,
-  AdapterInputError,
-  AdapterResponseInvalidError,
   Requester,
   util,
   Validator,
@@ -44,7 +42,6 @@ const inputParameters: InputParameters<TInputParameters> = {
   },
   operand1Input: {
     required: true,
-    type: 'object',
     description: 'The payload to send to the operand1 sources',
   },
   operand2Sources: {
@@ -60,7 +57,6 @@ const inputParameters: InputParameters<TInputParameters> = {
   },
   operand2Input: {
     required: true,
-    type: 'object',
     description: 'The payload to send to the operand2 sources',
   },
   operation: {
@@ -73,6 +69,14 @@ const inputParameters: InputParameters<TInputParameters> = {
 
 export const execute: ExecuteWithConfig<Config> = (input, _, config) => {
   const validator = new Validator(input, inputParameters)
+  validator.validated.data.operand1Input = validateInputPayload(
+    validator.validated.data.operand1Input,
+    'operand1Input',
+  )
+  validator.validated.data.operand2Input = validateInputPayload(
+    validator.validated.data.operand2Input,
+    'operand2Input',
+  )
   return executeComputedPrice(validator.validated.id, validator.validated.data, config)
 }
 
@@ -221,4 +225,22 @@ export const median = (values: number[]): Decimal => {
   const half = Math.floor(values.length / 2)
   if (values.length % 2) return new Decimal(values[half])
   return new Decimal(values[half - 1] + values[half]).div(2)
+}
+
+export const validateInputPayload = (input: string | object, inputName: string) => {
+  if (typeof input === 'string') {
+    try {
+      return JSON.parse(input)
+    } catch (e) {
+      throw new AdapterInputError({
+        statusCode: 400,
+        message: `Input payload for "${inputName}" is not valid JSON.`,
+      })
+    }
+  } else if (typeof input === 'object' && input !== null && !Array.isArray(input)) return input
+
+  throw new AdapterInputError({
+    statusCode: 400,
+    message: `Invalid input payload type for "${inputName}", expected JSON string or object`,
+  })
 }