+++ packages/sources/layer2-sequencer-health/src/evm.ts
@@ -99,6 +99,12 @@ export const sendEVMDummyTransaction = async (
       gasPrice: 0,
       to: wallet.address,
     },
+    [Networks.Xlayer]: {
+      value: 0,
+      gasLimit: 0,
+      gasPrice: 0,
+      to: wallet.address,
+    },
   }
   await race<ethers.providers.TransactionResponse>({
     timeout,
@@ -153,6 +159,10 @@ const lastSeenBlock: Record<EVMNetworks, { block: number; timestamp: number }> =
     timestamp: 0,
   },
   [Networks.Celo]: {
+    block: 0,
+    timestamp: 0,
+  },
+  [Networks.Xlayer]: {
     block: 0,
     timestamp: 0,
   },