+++ packages/sources/solana-functions/src/endpoint/index.ts
@@ -1,3 +1,4 @@
 export { endpoint as anchorData } from './anchor-data'
+export { endpoint as bufferLayout } from './buffer-layout'
 export { endpoint as eusxPrice } from './eusx-price'
 export { endpoint as sanctumInfinity } from './sanctum-infinity'