+++ packages/sources/streamex/src/index.ts
@@ -1,0 +1,20 @@
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { Adapter } from '@chainlink/external-adapter-framework/adapter'
+import { config } from './config'
+import { reserve } from './endpoint'
+
+export const adapter = new Adapter({
+  defaultEndpoint: reserve.name,
+  name: 'STREAMEX',
+  config,
+  endpoints: [reserve],
+  rateLimiting: {
+    tiers: {
+      default: {
+        rateLimit1m: 10,
+      },
+    },
+  },
+})
+
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)