+++ packages/sources/solana-functions/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @chainlink/solana-functions-adapter
 
+## 1.2.0
+
+### Minor Changes
+
+- [#4156](https://github.com/smartcontractkit/external-adapters-js/pull/4156) [`8d4db8a`](https://github.com/smartcontractkit/external-adapters-js/commit/8d4db8a0b33918240e8d8a2e1bbf560def31837a) Thanks [@dskloetc](https://github.com/dskloetc)! - Add buffer-layout endpoint
+
+## 1.1.2
+
+### Patch Changes
+
+- [#4131](https://github.com/smartcontractkit/external-adapters-js/pull/4131) [`4c94fd9`](https://github.com/smartcontractkit/external-adapters-js/commit/4c94fd916bd8b57898ae61cde0888a09fc543a6e) Thanks [@mxiao-cll](https://github.com/mxiao-cll)! - Bumped framework version to fix a websocket connection issue
+
+## 1.1.1
+
+### Patch Changes
+
+- [#4118](https://github.com/smartcontractkit/external-adapters-js/pull/4118) [`1136034`](https://github.com/smartcontractkit/external-adapters-js/commit/113603435a15a9f760ba1d16c4d70822dc358b75) Thanks [@mxiao-cll](https://github.com/mxiao-cll)! - Bumped framework version
+
+- [#4115](https://github.com/smartcontractkit/external-adapters-js/pull/4115) [`52dd8a4`](https://github.com/smartcontractkit/external-adapters-js/commit/52dd8a4fc30f2eed057078ff6c145eda50f6fba7) Thanks [@mxiao-cll](https://github.com/mxiao-cll)! - Code refactor
+
 ## 1.1.0
 
 ### Minor Changes